The two lines had fully diverged: merge-base was the initial commit, 27
commits on the feature side against 51 on origin/main, 80 files modified on
both. They are not rival products — origin/main already merged this line's
base via PR #26 (as c22ccf1), then moved forward with dependency bumps, CI
hardening and a ruff autofix pass.

Reconciling now rather than after the rename: Phase 2 of the rename moves
~150 paths under apps/backend/src/, and a rename-vs-modify merge across 80
core files is the shape that silently drops commits.

11 conflicts, resolved as follows.

Took this line's version, because it is semantically newer and origin/main's
side predates Phase 8.1/8.2:
- config.py — GRACE_* aliases (Phase 8.2 Part 5); origin/main was AXIOM_-only
- gateway/vault.py — origin/main still called aes_vault.decrypt(..., _kek()),
  which would have reintroduced the exact defect Phase 8.2 removed: the
  evidence key opening user credentials
- mcp/auth.py — GRACE_API_KEY with the AXIOM_API_KEY fallback
- agent-form.tsx — provider-aware model select; origin/main had free text
- empty-state.tsx — the four-step quickstart including CREDENTIAL
- MCP.md, decisions.md — verified this line is a strict superset (only a
  trailing separator differs before ADR-030)

Hand-merged rather than picking a side:
- routers/v1/events.py — origin/main's contextlib.suppress form with this
  line's aclose(); close() is the deprecated alias
- test_v1_events_sse.py — kept the root-cause fix (ASGITransport buffers the
  whole response, so stream=True never returns against a `while True`
  generator; the streaming tests now run against a real socket with bounded
  reads) AND origin/main's pytest-timeout safety net, which survived in
  pyproject. One fixes the hang, the other catches the next one.

Lockfiles regenerated from origin/main's as the base, then reconciled against
the merged manifests. This pulls in origin/main's dependency work, which is
substantial and would otherwise have been lost: Next 14.2.35 -> 16.3.6, plus
vite, vitest, eslint-config-next, postcss and the SDK bumps.

Verified post-merge: no _kek() in the gateway, every Phase 8.1/8.2/9.1 and
licensing artefact still present, and alembic reports a single head.

+++ apps/backend/src/axiom/services/governance/receipt_duration.py
@@ -16,7 +16,7 @@ def compute_receipt_duration_ms(receipt: GovernanceReceipt) -> int | None:
     if receipt.status != "sealed" or receipt.sealed_at is None:
         return None
     delta = receipt.sealed_at - receipt.created_at
-    ms = int(round(delta.total_seconds() * 1000))
+    ms = round(delta.total_seconds() * 1000)
     if ms < 0 or ms > _MAX_PLAUSIBLE_MS:
         logger.warning(
             "governance.receipt.duration_ms_clamped",
